// src/api/model/apiProcessingAction.ts
export interface ApiProcessingAction { 
    /**
     * Entity id
     */
    id?: number;
    /**
     * Processing action name
     */
    name?: string;
    /**
     * Processing action description
     */
    description?: string;
    /**
     * Processing action language
     */
    language?: ApiProcessingAction.LanguageEnum;
    /**
     * Sort order number. Lower number means first
     */
    sortOrder?: number;
    /**
     * Processing action prefix
     */
    prefix?: string;
    /**
     * The estimated output quantity per unit for this processing action
     */
    estimatedOutputQuantityPerUnit?: number;
    /**
     * Processing action public timeline label
     */
    publicTimelineLabel?: string;
    /**
     * Processing action public timeline location
     */
    publicTimelineLocation?: string;
    company?: ApiCompanyBase;
    inputSemiProduct?: ApiSemiProduct;
    /**
     * List of supported output semi-products
     */
    outputSemiProducts?: Array<ApiProcessingActionOutputSemiProduct>;
    inputFinalProduct?: ApiFinalProduct;
    outputFinalProduct?: ApiFinalProduct;
    /**
     * Processing action repacked outputs when using output final product
     */
    repackedOutputFinalProducts?: boolean;
    /**
     * Processing action maximum output weight when repacked outputs for final product is set to 'true'
     */
    maxOutputWeight?: number;
    qrCodeForFinalProduct?: ApiFinalProduct;
    /**
     * Processing action type
     */
    type?: ApiProcessingAction.TypeEnum;
    /**
     * Processing action public timeline icon type
     */
    publicTimelineIconType?: ApiProcessingAction.PublicTimelineIconTypeEnum;
    /**
     * Denoting if this processing action is Transfer or Quote of a Final product
     */
    finalProductAction?: boolean;
    /**
     * List of value chains for this processing action
     */
    valueChains?: Array<ApiValueChain>;
    /**
     * Processing action required document types
     */
    requiredDocumentTypes?: Array<ApiProcessingEvidenceType>;
    /**
     * Processing action required evidence fields
     */
    requiredEvidenceFields?: Array<ApiProcessingEvidenceField>;
    /**
     * List of facilities where this processing starts
     */
    supportedFacilities?: Array<ApiFacility>;
    /**
     * Processing action translations
     */
    translations?: Array<ApiProcessingActionTranslation>;
}

/**
 * Namespace for property- and property-value-enumerations of ApiProcessingAction.
 */
export namespace ApiProcessingAction {
// ...
    export enum LanguageEnum {
        EN = 'EN',
        DE = 'DE',
        RW = 'RW',
        FR = 'FR',
        ES = 'ES'
    }

    /**
     * All possible values of type.
     */
    export enum TypeEnum {
        PROCESSING = 'PROCESSING',
        FINALPROCESSING = 'FINAL_PROCESSING',
        SHIPMENT = 'SHIPMENT',
        TRANSFER = 'TRANSFER',
        GENERATEQRCODE = 'GENERATE_QR_CODE'
    }

    /**
     * All possible values of publicTimelineIconType.
     */
    export enum PublicTimelineIconTypeEnum {
        SHIP = 'SHIP',
        LEAF = 'LEAF',
        WAREHOUSE = 'WAREHOUSE',
        QRCODE = 'QRCODE',
        OTHER = 'OTHER'
    }
// ...
}